simplify resource loader
