+++ components/loader/Cargo.toml
@@ -7,5 +7,4 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-relative-path = "1.4.0"
 url = { path = "../url" }
